+++ client/app/modules/navbar/contact.service.js
@@ -1,0 +1,13 @@
+'use strict';
+
+export class ContactService {
+  /*@ngInject*/
+  constructor($http) {
+    this.$http = $http;
+  }
+
+  async messageSend(message) {
+    const response = await this.$http.post('/api/message', message);
+    return response;
+  }
+}