+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 # Product Backlog
-* [High] Convert schedulemanager, and navbar to modules (split out modal controllers and services)
-* [High] Create ContactService.sendMessage() for navbar.component.js
+* [High] Convert schedulemanager module (split out modal controllers and services)
 * [High] Create ScheduleService for schedulemanager.component.js (and add LocationService, TeachersService, ClassesService)
 * [High] Extend User service, forgotPassword() for login.component.js
 * [High] Create HistoryService for usermanager.component.js (and add LocationService, TeacherService, ClassesService) and merge AttendanceService into it